Pagination Added

+++ controllers/admin.js
@@ -1,3 +1,5 @@
+const fileHelper = require('../util/file')
+
 const {validationResult} = require('express-validator');
 
 const Product = require('../models/product');
@@ -158,6 +160,7 @@ exports.postEditProduct = (req,res,next) => {
             product.price = updatedPrice;
             product.description = updatedDescription;
             if (image) {
+                fileHelper.deleteFile(product.imageUrl);
                 product.imageUrl = image.path;
             }
             return product.save().then(result => {
@@ -173,13 +176,17 @@ exports.postEditProduct = (req,res,next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({_id: prodId, userId: req.user._id})
-        .then(() => {
-            console.log("Product Removed");
-            res.redirect('/admin/products');
-        }).catch(err => {
-            const error = new Error(err);
-            error.httpStatusCode = 500;
-            return next(error);
-        });
+    Product.findById(prodId).then(prod => {
+        if (!prod){
+            return next(new Error('Product not found'));
+        }
+        fileHelper.deleteFile(prod.imageUrl);
+        return Product.deleteOne({_id: prodId, userId: req.user._id});
+    }).then(() => {
+        res.redirect('/admin/products');
+    }).catch(err => {
+        const error = new Error(err);
+        error.httpStatusCode = 500;
+        return next(error);
+    });
 }
